fix(auth): wrap login page search params in suspense
